requirements.txt
matplotlib==3.7.1
matplotlib-inline==0.1.2
numpy==1.24.2
opencv-python==4.7.0.72
pandas==1.5.3
pickleshare==0.7.5
Pillow==9.5.0
plotly==5.14.1
pytesseract==0.3.10
scikit-learn==1.2.2
seaborn==0.12.2
streamlit==1.21.0
streamlit-option-menu==0.3.2